requirements.txt
# Core Web Framework
streamlit==1.26.1
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-multipart==0.0.6
requests==2.31.0

# ML & Deep Learning
torch>=2.0.0
transformers==4.40.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
timm>=0.9.0
efficientnet-pytorch>=0.7.1

# Image Processing
Pillow==10.0.0
opencv-python>=4.8.0
mediapipe==0.10.0
face-recognition>=1.3.0

# LLM & RAG
openai==1.0.0
google-generativeai>=0.3.0
langchain>=0.1.0
langchain-openai>=0.0.5

# Utilities
pydantic>=2.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
feedparser>=6.0.10

# Authentication
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
PyJWT>=2.8.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
